update-maintainer: Find debian/control from inside debian/
In addition to looking for ./debian in the current directory, scan up
the path in the case of running update-maintain from inside the debian/
directory.  This scans up to a maximum of 6 levels.

The --help text is adjusted to identify the detected path if one was
found, defaulting to './debian' otherwise.

LP: #1885233
Signed-off-by: Bryce Harrington <bryce@bryceharrington.org>