ubuntu-build: fix non batch mode errors
The current version is not working

```
$ ubuntu-build librsync resolute retry

Traceback (most recent call last):
  File "/usr/lib/python3/dist-packages/lazr/restfulclient/resource.py", line 354, in __getattr__
    return self.lp_get_parameter(attr)
           ~~~~~~~~~~~~~~~~~~~~~^^^^^^
  File "/usr/lib/python3/dist-packages/lazr/restfulclient/resource.py", line 249, in lp_get_parameter
    raise KeyError("No such parameter: %s" % param_name)
KeyError: 'No such parameter: getComponent'
```

The way launchpadlib is used was changed in 010af53 but non batch mode
was not correctly updated.

Also tweak the way the release is computed for distroseries to be able
to retry a package in e.g "resolute-proposed".